COM-1738 - fix hitSlop
