+++ src/components/IconButton/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  button: {
+    zIndex: 100,
+  },
+});
+
+export default styles;
